+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-core`
 
+## 1.17.0.0
+
+* Change superclass of `Crypto` from `KESAlgorithm` to `UnsoundPureKESAlgorithm`
+
 ## 1.16.0.0
 
 * Add `PlutusDebugOverrides` argument to `debugPlutus`